Option to use posix exit code upon fatal signal

+++ bin/mocha.js
@@ -10,6 +10,7 @@
  * @private
  */
 
+const os = require('os');
 const {loadOptions} = require('../lib/cli/options');
 const {
   unparseNodeFlags,
@@ -109,7 +110,12 @@ if (mochaArgs['node-option'] || Object.keys(nodeArgs).length || hasInspect) {
   proc.on('exit', (code, signal) => {
     process.on('exit', () => {
       if (signal) {
-        process.kill(process.pid, signal);
+        if (mochaArgs['posix-exit-codes'] === true) {
+          process.exitCode = 128 + os.constants.signals[signal];
+          process.exit(process.exitCode);
+        } else {
+          process.kill(process.pid, signal);
+        }
       } else {
         process.exit(code);
       }
@@ -126,7 +132,7 @@ if (mochaArgs['node-option'] || Object.keys(nodeArgs).length || hasInspect) {
     // be needed.
     if (!args.parallel || args.jobs < 2) {
       // win32 does not support SIGTERM, so use next best thing.
-      if (require('os').platform() === 'win32') {
+      if (os.platform() === 'win32') {
         proc.kill('SIGKILL');
       } else {
         // using SIGKILL won't cleanly close the output streams, which can result
